% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whitelists.r
\name{get.default.whitelist}
\alias{get.default.whitelist}
\title{Get a stored default whitelist for a specified package}
\usage{
get.default.whitelist(pkg = "base", color = c("white", "black")[1],
  funvar = c("fun", "var")[1], group = "", return.type = c("vector",
  "yaml", "list")[1])
}
\arguments{
\item{pkg}{the package name}

\item{color}{"white" for whitelist and "black" for blacklist}

\item{funvar}{"fun" for function list "var" for variable list}

\item{group}{possibly the name of a list group}

\item{return.type}{"vector" names of whitelisted functiin, "yaml" the yaml source}
}
\description{
Get a stored default whitelist for a specified package
}

